Create README.md
Further to a large study I undertook in 2019 regarding global Income Inequality (Gini Index) and Life Expectancy (in R),
this study is an exploration of global Health Expenditure (USD) per capita, and its relation to the "Health Development Index" (United Nations).

Extensive reading about the corrosive effects of income inequality (Gini Index) highlighted a body of research, which proposed that a nation's health outcomes were a direct effect of that nation's expenditure on health per capita, irrespective of wider social challenges (like wealth inequality).  In other words, if a nation wants good health outcomes, they must pay for them! This is an exploratory study of this intricate and fascinating area.

Many thanks to the many Githubers, Kagglers and Youtubers who have provided me with so much guidance and inspiration.